requirements.txt
setuptools~=65.5.0
packaging~=23.1
pytest~=7.3.2
requests~=2.31.0
pydantic~=1.10.9
PyYAML~=6.0
python-dateutil~=2.8.2